Use float32 everywhere

+++ dsd.go
@@ -8,14 +8,17 @@ var (
 	ErrNotImplemented = errors.New("dsd: not implemented")
 )
 
+// Decoder transforms encoded bits to float32 samples.
 type Decoder interface {
 	Decode(bits []byte) ([]float32, error)
 }
 
+// Encoder transforms float32 samples to encoded bits.
 type Encoder interface {
-	Encode([]int32) ([]byte, error)
+	Encode([]float32) ([]byte, error)
 }
 
+// VoiceStream is a complete codec with a decode and an encoder.
 type VoiceStream interface {
 	Decoder
 	Encoder
